move HandlerCommandValidationFailure from IEventSourcedAggregate<T> to EventSourcedAggregate
